+++ src/controller.ohlc.js
@@ -2,31 +2,29 @@
 
 module.exports = function(Chart) {
 
-
-	Chart.defaults.ohlc = Chart.defaults.candlestick;
+	Chart.defaults.ohlc = Chart.defaults.financial;
 	
-	
-	Chart.controllers.ohlc = Chart.controllers.candlestick.extend({
+	Chart.controllers.ohlc = Chart.controllers.financial.extend({
 		
 		dataElementType: Chart.elements.ohlc,
 		
-		updateElement: function(candle, index, reset) {
+		updateElement: function(element, index, reset) {
 			var me = this;
 			var meta = me.getMeta();
 			var dataset = me.getDataset();
-			candle._xScale = me.getScaleForId(meta.xAxisID);
-			candle._yScale = me.getScaleForId(meta.yAxisID);
-			candle._datasetIndex = me.index;
-			candle._index = index;
-			candle._model = {
+			element._xScale = me.getScaleForId(meta.xAxisID);
+			element._yScale = me.getScaleForId(meta.yAxisID);
+			element._datasetIndex = me.index;
+			element._index = index;
+			element._model = {
 				datasetLabel: dataset.label || '',
 				lineWidth: dataset.lineWidth,
 				armLength: dataset.armLength,
 				armLengthRatio: dataset.armLengthRatio,
 				color: dataset.color,
 			};
-			me.updateElementGeometry(candle, index, reset);
-			candle.pivot();
+			me.updateElementGeometry(element, index, reset);
+			element.pivot();
 		},
 		
 	});
