+++ src/element.candlestick.js
@@ -2,50 +2,20 @@
 
 module.exports = function(Chart) {
 
-	var helpers = Chart.helpers,
-		globalOpts = Chart.defaults.global;
+	var helpers = Chart.helpers;
+	var globalOpts = Chart.defaults.global;
 
-	globalOpts.elements.candlestick = {
-		upCandleColor: "rgba(80, 160, 115, 1)",
-		downCandleColor: "rgba(215, 85, 65, 1)",
-		outlineCandleColor: "rgba(90, 90, 90, 1)",
-		outlineCandleWidth: 1,
-	};
+	globalOpts.elements.candlestick = Object.assign(globalOpts.elements.financial, {
+		color: Object.assign(globalOpts.elements.financial.color, {
+			outline: globalOpts.elements.financial.color.linear,
+		}),
+		outlineWidth: 1,
+	});
 
-	function isVertical(bar) {
-		return bar._view.width !== undefined;
-	}
-
-	/**
-	 * Helper function to get the bounds of the candle
-	 * @private
-	 * @param bar {Chart.Element.Candlestick} the bar
-	 * @return {Bounds} bounds of the bar
-	 */
-	function getBarBounds(candle) {
-		var vm = candle._view;
-		var x1, x2, y1, y2;
-
-		var halfWidth = vm.width / 2;
-		x1 = vm.x - halfWidth;
-		x2 = vm.x + halfWidth;
-		y1 = vm.candle.h;
-		y2 = vm.candle.l;
-
-
-		return {
-			left: x1,
-			top: y1,
-			right: x2,
-			bottom: y2
-		};
-	}
-
-	Chart.elements.Candlestick = Chart.Element.extend({
+	Chart.elements.candlestick = Chart.elements.financial.extend({
 		draw: function() {
 			var ctx = this._chart.ctx;
 			var vm = this._view;
-
 
 			var x = vm.x;
 			var o = vm.candle.o;
@@ -53,14 +23,14 @@ module.exports = function(Chart) {
 			var l = vm.candle.l;
 			var c = vm.candle.c;
 
-			ctx.strokeStyle = helpers.getValueOrDefault(vm.outlineCandleColor, globalOpts.elements.candlestick.outlineCandleColor);
-			ctx.lineWidth = helpers.getValueOrDefault(vm.outlineCandleWidth, globalOpts.elements.candlestick.outlineCandleWidth);
+			ctx.strokeStyle = helpers.getValueOrDefault(vm.color?vm.color.outline:undefined, globalOpts.elements.candlestick.color.outline);
+			ctx.lineWidth = helpers.getValueOrDefault(vm.outlineWidth, globalOpts.elements.candlestick.outlineWidth);
 			if (c < o) {
-				ctx.fillStyle = helpers.getValueOrDefault(vm.upCandleColor, globalOpts.elements.candlestick.upCandleColor);
+				ctx.fillStyle = helpers.getValueOrDefault(vm.color?vm.color.up:undefined, globalOpts.elements.candlestick.color.up);
 			} else if (c > o) {
-				ctx.fillStyle = helpers.getValueOrDefault(vm.downCandleColor, globalOpts.elements.candlestick.downCandleColor);
+				ctx.fillStyle = helpers.getValueOrDefault(vm.color?vm.color.down:undefined, globalOpts.elements.candlestick.color.down);
 			} else {
-				ctx.fillStyle = helpers.getValueOrDefault(vm.outlineCandleColor, globalOpts.elements.candlestick.outlineCandleColor);
+				ctx.fillStyle = helpers.getValueOrDefault(vm.color?vm.color.linear:undefined, globalOpts.elements.candlestick.color.linear);
 			}
 
 			ctx.beginPath();
@@ -73,67 +43,5 @@ module.exports = function(Chart) {
 			ctx.strokeRect(x - vm.width / 2, c, vm.width, o - c);
 			ctx.closePath();
 		},
-		height: function() {
-			var vm = this._view;
-			return vm.base - vm.y;
-		},
-		inRange: function(mouseX, mouseY) {
-			var inRange = false;
-
-			if (this._view) {
-				var bounds = getBarBounds(this);
-				inRange = mouseX >= bounds.left && mouseX <= bounds.right && mouseY >= bounds.top && mouseY <= bounds.bottom;
-			}
-
-			return inRange;
-		},
-		inLabelRange: function(mouseX, mouseY) {
-			var me = this;
-			if (!me._view) {
-				return false;
-			}
-
-			var inRange = false;
-			var bounds = getBarBounds(me);
-
-			if (isVertical(me)) {
-				inRange = mouseX >= bounds.left && mouseX <= bounds.right;
-			} else {
-				inRange = mouseY >= bounds.top && mouseY <= bounds.bottom;
-			}
-
-			return inRange;
-		},
-		inXRange: function(mouseX) {
-			var bounds = getBarBounds(this);
-			return mouseX >= bounds.left && mouseX <= bounds.right;
-		},
-		inYRange: function(mouseY) {
-			var bounds = getBarBounds(this);
-			return mouseY >= bounds.top && mouseY <= bounds.bottom;
-		},
-		getCenterPoint: function() {
-			var vm = this._view;
-			var x, y;
-
-			var halfWidth = vm.width / 2;
-			x = vm.x - halfWidth;
-			y = (vm.candle.h + vm.candle.l) / 2;
-
-			return { x: x, y: y };
-		},
-		getArea: function() {
-			var vm = this._view;
-			return vm.width * Math.abs(vm.y - vm.base);
-		},
-		tooltipPosition: function() {
-			var vm = this._view;
-			return {
-				x: vm.x,
-				y: (vm.candle.h + vm.candle.l) / 2
-			};
-		}
 	});
-
 };
-
